--- a/network/p2p/blob_service.go
+++ b/network/p2p/blob_service.go
@@ -26,6 +26,8 @@ import (
 	"github.com/onflow/flow-go/module/irrecoverable"
 	"github.com/onflow/flow-go/module/metrics"
 	"github.com/onflow/flow-go/network"
+
+	ipld "github.com/ipfs/go-ipld-format"
 )
 
 type blobService struct {
@@ -166,7 +168,7 @@ func (bs *blobService) TriggerReprovide(ctx context.Context) error {
 
 func (bs *blobService) GetBlob(ctx context.Context, c cid.Cid) (blobs.Blob, error) {
 	blob, err := bs.blockService.GetBlock(ctx, c)
-	if err == blockservice.ErrNotFound {
+	if ipld.IsNotFound(err) {
 		return nil, network.ErrBlobNotFound
 	}