--- a/network/test/middleware_test.go
+++ b/network/test/middleware_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	swarm "github.com/libp2p/go-libp2p-swarm"
+	"github.com/libp2p/go-libp2p/p2p/net/swarm"
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"