* https://github.com/libp2p/go-libp2p-swarm is deprecated.
* https://github.com/libp2p/go-libp2p-discovery is deprecated.
* https://github.com/libp2p/go-tcp-transport is deprecated.
* `blockstore.ErrNotFound` is no longer exported exported and replaced with `ipld.IsNotFound(err)`.
* `go vet` [now complains about wrong usage](https://tip.golang.org/doc/go1.19#vet) of `errors.As` which we have plenty 

As a tiny bonus, a 3% perf improvement based on the ComputeBlock benchmark:
```
$ benchstat old new   
name                             old time/op    new time/op    delta
ComputeBlock/16/cols/128/txes-8     2.09s ± 1%     2.03s ± 2%  -2.99%  (p=0.000 n=9+10)

name                             old us/tx      new us/tx      delta
ComputeBlock/16/cols/128/txes-8     1.02k ± 1%     0.99k ± 2%  -2.98%  (p=0.000 n=9+10)

name                             old alloc/op   new alloc/op   delta
ComputeBlock/16/cols/128/txes-8    1.88GB ± 1%    1.87GB ± 0%    ~     (p=0.370 n=9+8)

name                             old allocs/op  new allocs/op  delta
ComputeBlock/16/cols/128/txes-8     29.3M ± 0%     29.2M ± 0%  -0.13%  (p=0.000 n=8+8)
```