@SaveTheRbtz 2022-08-16
@tarakby seems like `github.com/btcsuite/btcd` is no longer developed and crypto would need to migrate to `github.com/btcsuite/btcd/btcec/v2` at some point.