@tarakby 2022-08-16
>  seems like github.com/btcsuite/btcd is no longer developed and ./crypto/ would need to migrate to github.com/btcsuite/btcd/btcec/v2 at some point.

There is a community PR with the upgrade, I'll follow up on that.